Implement multiple color stops

+++ src/main/java/hu/awm/srtm/helper/Gradient.java
@@ -3,5 +3,5 @@ package hu.awm.srtm.helper;
 import java.awt.*;
 
 public interface Gradient {
-	Color heightToColor(int height);
+	Color heightToColor(double height);
 }
